refactor(mcp): extract update_backlog_item handler out of the mcp-tools god-file
Module Size Guard: the BI-87A8BF2A additions pushed mcp-tools.ts (16.5k LOC) past
its baseline. Rather than ratchet the baseline up, extract the whole
update_backlog_item case into apps/web/lib/mcp-handlers/update-backlog-item.ts and
delegate — mcp-tools.ts shrinks below baseline (16513 <= 16517) and the handler is
now unit-addressable. Behaviour unchanged; result nests portfolioId under data
per ToolResult.
